fix: minor comments
