+++ libs/application/templates/directorate-of-equality/equality-report/src/lib/messages.ts
@@ -83,7 +83,7 @@ export const messages = {
     },
   }),
 
-  // Forsendur
+  //
   prerequisites: {
     errors: defineMessages({
       approveExternalData: {
